History Services
Mongo repository implementation and testing

+++ common-web/src/main/java/com/pabloam/microservices/converter/common/ConvertedResponse.java
@@ -22,7 +22,7 @@ public interface ConvertedResponse {
 	 * 
 	 * @return
 	 */
-	public Map<String, Float> getQuotes();
+	public Map<String, Double> getQuotes();
 
 	/**
 	 * The timestamp returned from the provider
